@@ -11,6 +11,7 @@
   "files": [
     "dist"
   ],
+  "type": "module",
   "main": "./dist/mui-tel-input.es.js",
   "types": "./dist/index.d.ts",
   "exports": {
@@ -39,7 +40,7 @@
   ],
   "scripts": {
     "build": "vite build",
-    "lint": "tsc && eslint . --ext .js,.jsx,.ts,.tsx",
+    "lint": "npx tsc && eslint",
     "lint:fix": "npm run lint -- --fix",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
@@ -60,7 +61,7 @@
   "peerDependencies": {
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
-    "@mui/material": "^6.0.0",
+    "@mui/material": "^7.0.0",
     "@types/react": "^18.0.0 || ^19.0.0",
     "react": "^18.0.0 || ^19.0.0",
     "react-dom": "^18.0.0 || ^19.0.0"
@@ -78,12 +79,11 @@
     "@babel/core": "^7.26.10",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@mui/material": "^6.4.6",
+    "@mui/material": "^7.0.2",
     "@storybook/addon-actions": "^8.6.12",
     "@storybook/addon-essentials": "^8.6.12",
     "@storybook/addon-interactions": "^8.6.12",
     "@storybook/addon-links": "^8.6.12",
-    "@storybook/react": "^8.6.12",
     "@storybook/react-vite": "^8.6.12",
     "@storybook/test": "^8.6.12",
     "@testing-library/dom": "^10.4.0",
@@ -92,15 +92,15 @@
     "@testing-library/user-event": "^14.6.1",
     "@types/react": "^19.1.1",
     "@types/react-dom": "^19.1.2",
-    "@viclafouch/eslint-config-viclafouch": "4.16.0",
+    "@viclafouch/eslint-config-viclafouch": "4.22.0",
     "@vitejs/plugin-react": "^4.3.4",
-    "babel-loader": "^9.2.1",
-    "eslint": "^8.56.0",
+    "babel-loader": "^10.0.0",
+    "eslint": "^9.24.0",
     "husky": "^9.1.7",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
     "jsdom": "^26.1.0",
     "prettier": "^3.5.3",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "standard-version": "^9.5.0",
     "storybook": "^8.6.12",
